+++ src/schema/mutation/addApplication.mutation.ts
@@ -17,7 +17,7 @@ export default {
   async resolve(parent, args, context) {
     const user = context.user;
     if (!user) {
-      throw new GraphQLError(context.i18next.t('errors.userNotLogged'));
+      throw new GraphQLError(context.i18next.t('common.errors.userNotLogged'));
     }
     const ability: AppAbility = user.ability;
     if (ability.can('create', 'Application')) {
@@ -46,7 +46,7 @@ export default {
         name: appName,
         //createdAt: new Date(),
         status: status.pending,
-        createdBy: user.id,
+        createdBy: user._id,
         permissions: {
           canSee: [],
           canUpdate: [],
@@ -100,7 +100,9 @@ export default {
       }
       return application;
     } else {
-      throw new GraphQLError(context.i18next.t('errors.permissionNotGranted'));
+      throw new GraphQLError(
+        context.i18next.t('common.errors.permissionNotGranted')
+      );
     }
   },
 };