+++ src/schema/mutation/seeNotification.mutation.ts
@@ -16,7 +16,7 @@ export default {
     // Authentication check
     const user = context.user;
     if (!user) {
-      throw new GraphQLError(context.i18next.t('errors.userNotLogged'));
+      throw new GraphQLError(context.i18next.t('common.errors.userNotLogged'));
     }
 
     const ability: AppAbility = context.user.ability;
@@ -24,7 +24,7 @@ export default {
       .where({ _id: args.id })
       .getFilter();
     return Notification.findOneAndUpdate(filters, {
-      $push: { seenBy: user.id },
+      $push: { seenBy: user._id },
     });
   },
 };