+++ src/schema/mutation/seeNotifications.mutation.ts
@@ -21,20 +21,20 @@ export default {
     // Authentication check
     const user = context.user;
     if (!user) {
-      throw new GraphQLError(context.i18next.t('errors.userNotLogged'));
+      throw new GraphQLError(context.i18next.t('common.errors.userNotLogged'));
     }
 
     const ability: AppAbility = context.user.ability;
     if (!args) {
       throw new GraphQLError(
-        context.i18next.t('errors.invalidSeeNotificationsArguments')
+        context.i18next.t('mutations.notification.see.errors.invalidArguments')
       );
     }
     const filters = Notification.accessibleBy(ability, 'update')
       .where({ _id: { $in: args.ids } })
       .getFilter();
     const result = await Notification.updateMany(filters, {
-      $push: { seenBy: user.id },
+      $push: { seenBy: user._id },
     });
     return result.ok === 1;
   },