+++ src/security/defineUserAbility.ts
@@ -307,7 +307,7 @@ export default function defineUserAbility(user: User | Client): AppAbility {
         )
         .flat(),
     },
-    seenBy: { $ne: user.id },
+    seenBy: { $ne: user._id },
   });
 
   /* ===