+++ src/utils/user/updateUserAttributes.ts
@@ -31,9 +31,12 @@ export const updateUserAttributes = async (
     isEmpty(settings.mapping)
   ) {
     logger.error(
-      i18next.t('errors.missingObjectParameters', {
-        object: 'user attributes settings',
-      })
+      i18next.t(
+        'utils.user.updateUserAttributes.errors.missingObjectParameters',
+        {
+          object: 'user attributes settings',
+        }
+      )
     );
     return false;
   }
@@ -44,7 +47,7 @@ export const updateUserAttributes = async (
     settings.apiConfiguration
   );
   if (apiConfiguration.authType === authType.serviceToService) {
-    token = await getDelegatedToken(apiConfiguration, user.id, upstreamToken);
+    token = await getDelegatedToken(apiConfiguration, user._id, upstreamToken);
   }
   // Pass it in headers
   const headers: any = token
@@ -60,14 +63,14 @@ export const updateUserAttributes = async (
       headers,
     });
   } catch (err) {
-    logger.error(i18next.t('errors.invalidAPI'), { stack: err.stack });
+    logger.error(i18next.t('common.errors.invalidAPI'), { stack: err.stack });
     return false;
   }
   let data: any;
   try {
     data = await res.json();
   } catch (err) {
-    logger.error(i18next.t('errors.authenticationTokenNotFound'), {
+    logger.error(i18next.t('common.errors.authenticationTokenNotFound'), {
       stack: err.stack,
     });
     return false;